+++ journeyapps-parser-schema/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @journeyapps/parser-schema
 
+## 8.2.5
+
+### Patch Changes
+
+- f1f492c: Fix null pointer when accessing `displayFormat` on `ObjectType`
+- Updated dependencies [d2e1526]
+  - @journeyapps/evaluator@7.0.5
+  - @journeyapps/parser-common@8.0.7
+
+## 8.2.4
+
+### Patch Changes
+
+- Updated dependencies [9abb3f9]
+  - @journeyapps/evaluator@7.0.4
+  - @journeyapps/parser-common@8.0.6
+
+## 8.2.3
+
+### Patch Changes
+
+- Updated dependencies [a68b965]
+  - @journeyapps/evaluator@7.0.3
+  - @journeyapps/parser-common@8.0.5
+
+## 8.2.2
+
+### Patch Changes
+
+- Updated dependencies [16bb578]
+  - @journeyapps/evaluator@7.0.2
+  - @journeyapps/parser-common@8.0.4
+
 ## 8.2.1
 
 ### Patch Changes