+++ journeyapps/CHANGELOG.md
@@ -1,5 +1,37 @@
 # journeyapps
 
+## 7.3.16
+
+### Patch Changes
+
+- Updated dependencies [fda55dc]
+  - @journeyapps/db@8.1.0
+
+## 7.3.15
+
+### Patch Changes
+
+- Updated dependencies [d2e1526]
+  - @journeyapps/db@8.0.14
+
+## 7.3.14
+
+### Patch Changes
+
+- @journeyapps/db@8.0.13
+
+## 7.3.13
+
+### Patch Changes
+
+- @journeyapps/db@8.0.12
+
+## 7.3.12
+
+### Patch Changes
+
+- @journeyapps/db@8.0.11
+
 ## 7.3.11
 
 ### Patch Changes